Split classes into files, make GetValuesInOrder() an extension method
